requirements.txt
fastapi
uvicorn
python-multipart
openai-whisper
faster-whisper
pydub
librosa==0.10.1
#numpy==1.26.4
pyloudnorm
soundfile
scipy==1.11.4
aiofiles
requests
faster-whisper  
transformers
torch
ctranslate2
spacy
huggingface_hub
sentencepiece
torchaudio
sqlalchemy
passlib[bcrypt]
python-jose
slowapi
pydantic-settings
alembic





